# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Nexa Panel HTTP Interface
  version: 0.1.0
  description: Initial control-plane contract for the Nexa Panel foundation.
servers:
  - url: /api/v1
paths:
...
  /auth/bootstrap:
    post:
      operationId: bootstrapAdministrator
      tags: [Identity]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "201":
          description: The first administrator and session were created.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordResponse"
        "409":
          description: An administrator already exists.
  /auth/login:
    post:
      operationId: login
      tags: [Identity]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Credentials"
      responses:
        "200":
          description: Credentials were accepted and a session was created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordResponse"
        "401":
          description: The credentials were not accepted.
        "429":
          description: Too many password attempts were made from the client address.
  /auth/mfa/enroll:
    post:
      operationId: enrollMFA
      tags: [Identity]
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Encrypted TOTP enrollment material for the pre-authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MFAEnrollment"
        "409":
          description: MFA is already enrolled.
  /auth/mfa/confirm:
    post:
      operationId: confirmMFA
      tags: [Identity]
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [code]
              properties:
                code:
                  $ref: "#/components/schemas/TOTPCode"
      responses:
        "200":
          description: Enrollment was confirmed; recovery codes are returned once.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MFAConfirmation"
        "401":
          description: The TOTP was invalid, expired, or previously accepted.
        "429":
          description: Too many verification attempts were made.
  /auth/mfa/verify:
    post:
      operationId: verifyMFA
      tags: [Identity]
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  additionalProperties: false
                  required: [code]
                  properties:
                    code:
                      $ref: "#/components/schemas/TOTPCode"
                - type: object
                  additionalProperties: false
                  required: [recoveryCode]
                  properties:
                    recoveryCode:
                      type: string
                      pattern: "^[A-Fa-f0-9]{4}(-[A-Fa-f0-9]{4}){3}$"
      responses:
        "200":
          description: The pre-authenticated session is now fully authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: The TOTP or recovery code was invalid or already consumed.
        "429":
          description: Too many verification attempts were made.
  /auth/session:
    get:
      operationId: getSession
      tags: [Identity]
      security:
        - sessionCookie: []
      responses:
        "200":
          description: The authenticated administrator.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: A valid session is required.
  /auth/logout:
    post:
      operationId: logout
      tags: [Identity]
      security:
        - sessionCookie: []
      responses:
        "204":
          description: The current session was revoked.
  /users:
    get:
      operationId: listUsers
      tags: [Users]
      x-permission: users.manage
      security:
        - sessionCookie: []
      responses:
        "200":
          description: All managed users with their developer site grants.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ManagedUser"
    post:
      operationId: createUser
      tags: [Users]
      x-permission: users.manage
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUser"
      responses:
        "201":
          description: The managed user was created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedUser"
        "409": {description: A user with this username already exists., content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "422": {description: The username or password or role is not allowed., content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
  /users/{id}:
    patch:
      operationId: updateUser
      tags: [Users]
      x-permission: users.manage
      security:
        - sessionCookie: []
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUser"
      responses:
        "200":
          description: The updated user; every live session of the user was revoked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManagedUser"
        "400": {description: Neither a role nor a password was provided., content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "404": {description: The user does not exist., content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "409": {description: The last administrator cannot lose the admin role., content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "422": {description: The role or password is not allowed., content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
    delete:
      operationId: deleteUser
      tags: [Users]
      x-permission: users.manage
      security:
        - sessionCookie: []
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      responses:
        "204":
          description: The user plus their sessions and site grants were removed.
        "404": {description: The user does not exist., content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "409": {description: The last administrator or your own account cannot be deleted., content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
  /users/{id}/sites:
    put:
      operationId: replaceUserSiteGrants
      tags: [Users]
      x-permission: users.manage
      security:
        - sessionCookie: []
      parameters:
        - {name: id, in: path, required: true, schema: {type: string}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSiteGrants"
      responses:
        "200":
          description: The developer's site grants were replaced with the deduplicated set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSiteGrants"
        "400": {description: Site grants apply to developer users only., content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "404": {description: The user does not exist., content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "422": {description: A granted site does not exist., content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
        "503": {description: Site grants are not configured on this control plane., content: {application/json: {schema: {$ref: "#/components/schemas/Error"}}}}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: nexa_session
...
  schemas:
...
    User:
      type: object
      required: [id, username, role]
      properties:
        id:
          type: string
        username:
          type: string
          minLength: 3
          maxLength: 64
        role:
          type: string
          enum: [viewer, operator, admin]
    UserResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"
    PasswordResponse:
      type: object
      required: [user, next]
      properties:
        user:
          $ref: "#/components/schemas/User"
        next:
          type: string
          enum: [mfa_enrollment, mfa_challenge]
...
    TOTPCode:
      type: string
      pattern: "^[0-9]{6}$"
    MFAEnrollment:
      type: object
      required: [secret, provisioningUri]
      properties:
        secret:
          type: string
        provisioningUri:
          type: string
          pattern: "^otpauth://totp/"
    MFAConfirmation:
      type: object
      required: [user, recoveryCodes]
      properties:
        user:
          $ref: "#/components/schemas/User"
        recoveryCodes:
          type: array
          minItems: 10
          maxItems: 10
          items:
            type: string
    Credentials:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
        password:
          type: string
          format: password
          minLength: 1
          maxLength: 1024
    Error:
      description: Shared error envelope returned by every non-2xx JSON response.
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        offset:
          type: integer
          format: int64
          minimum: 0
          description: Present on upload offset conflicts; the next byte offset the server expects.
    ManagedRole:
      type: string
      enum: [admin, operator, developer, viewer]
    ManagedUser:
      type: object
      required: [id, username, role, createdAt, mfaConfirmed, siteIds]
      properties:
        id: {type: string}
        username: {type: string, minLength: 3, maxLength: 64}
        role: {$ref: "#/components/schemas/ManagedRole"}
        createdAt: {type: string, format: date-time}
        lastLoginAt: {type: string, format: date-time}
        mfaConfirmed: {type: boolean}
        siteIds:
          type: array
          items: {type: string}
          description: Sites granted to a developer; empty for unscoped roles.
    CreateUser:
      type: object
      additionalProperties: false
      required: [username, password, role]
      properties:
        username: {type: string, pattern: "^[A-Za-z0-9][A-Za-z0-9._-]{2,63}$"}
        password: {type: string, format: password, minLength: 12, maxLength: 1024}
        role: {$ref: "#/components/schemas/ManagedRole"}
    UpdateUser:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        role: {$ref: "#/components/schemas/ManagedRole"}
        password: {type: string, format: password, minLength: 12, maxLength: 1024}
    UserSiteGrants:
      type: object
      additionalProperties: false
      required: [siteIds]
      properties:
        siteIds:
          type: array
          items: {type: string}
...